validity check and comments data transferred to presenter
